requirements.txt
Django==1.6.5
Pillow==2.5.0
argparse==1.2.1
setuptools==1.0
dj-database-url==0.3.0
dj-static==0.0.5
django-embed-video==0.11
django-secure==1.0
django-toolbelt==0.0.1
django-wysiwyg-redactor==0.3.9.1
easy-thumbnails==2.0.1
gunicorn==19.0.0
psycopg2==2.5.3
pystache==0.5.4
requests==2.3.0
static==1.0.2
wsgiref==0.1.2
